test: fixed pull-request tests
